+++ tests/test_debug.py
@@ -210,6 +210,11 @@ class DebugTraceTest(CoverageTest):
         vtuple = ast.literal_eval(pyversion.partition(":")[-1].strip())
         assert vtuple[:5] == sys.version_info
 
+    def test_debug_pytest(self) -> None:
+        out_text = self.f1_debug_output(["trace", "pytest"])
+        ctx = "tests/test_debug.py::DebugTraceTest::test_debug_pytest (call)"
+        assert f"Pytest context: {ctx}" in out_text
+
 
 def assert_good_debug_sys(out_text: str) -> None:
     """Assert that `str` is good output for debug=sys."""